Rename ExistingDartel (files, folders, classes) to RegisterDartel
