+++ clinica/pipelines/t1_volume_register_dartel/info.json
