+++ clinica/pipelines/t1_volume_register_dartel/t1_volume_register_dartel_utils.py
